refactor: Main 페이지 내에서 사용하는 컴포넌트 파일 위치 변경
