+++ src/pages/Main/index.tsx
@@ -1,8 +1,8 @@
 import MessageCard from '@/components/MessageCard'
 import SolidButton from '@/components/SolidButton'
-import TopButton from '@/components/TopButton'
-import Header from '@/containers/Main/Header'
-import NoticeSection from '@/containers/Main/NoticeSection'
+import TopButton from './components/TopButton'
+import Header from './components/Header'
+import NoticeSection from './components/NoticeSection'
 import useBodyBackgroundColor from '@/hooks/useBodyBackgroundColor'
 import { useMainData } from '@/hooks/useMainData'
 import { useGetMessages } from '@/hooks/useMessage'
